requirements.txt
appdirs==1.4.3
bcrypt==3.1.3
cffi==1.10.0
Django==1.11
django-smartfields==1.0.9
olefile==0.44
packaging==16.8
Pillow==4.1.0
pycparser==2.17
pyparsing==2.2.0
pytz==2017.2
six==1.10.0
